refactored storage functions to handle generic JSON added first interaction with filter_rules added models for already retweeted tweets

+++ src/models/filter_query.rs
@@ -1,6 +1,6 @@
-#[derive(Serialize, Deserialize, Clone)]
+#[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct FilterQuery {
     pub q: String,
-    pub count: u8,
+    pub count: u32,
     pub result_type: String,
 }
